Split into modules instead of monolithic file as part of GHI #2
